🌟feat: refactor more disruption components for reusability
